DermaCareHacks-master/requirements.txt
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
Flask==1.1.2
google-api-core==1.17.0
google-auth==1.14.3
google-cloud-automl==0.10.0
googleapis-common-protos==1.51.0
grpcio==1.29.0
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
protobuf==3.12.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2020.1
requests==2.23.0
rsa==4.0
six==1.14.0
urllib3==1.25.9
Werkzeug==1.0.1
